rfapp Refactored the AbstractApplication class
The AbstractApplication class was part of a mini framework.
I moved all the code into the TurtleApp.
In the process I removed the Applicationable interface
and ApplicationStandard.
The code should be easier to follow now.
